fix(egui): renderモジュールに分離
レンダリング情報はアプリケーション内共通化をする
